Fix wrong type in stored trivial 3d biome structures.

+++ chunky/src/java/se/llbit/chunky/renderer/scene/biome/Trivial3dBiomeStructureImpl.java
@@ -20,7 +20,7 @@ public class Trivial3dBiomeStructureImpl extends Position3d2ReferencePackedArray
       XYZTriple key = entry.getKey();
       out.writeInt(key.x);
       out.writeInt(key.y);
-      out.writeLong(key.z);
+      out.writeInt(key.z);
       Object[] value = entry.getValue();
 
       BitSet presentValues = new BitSet(value.length);
